ndg: allow templating footer & navbar; match arbitrary templates with filenames
This addresses the lack of customization for specific components.
Namely, we now allow for the footer and the navbar to be templated
without having to replace the entire default template. This should
reduce the severity of having to vendor an entire template when all you
need is a basic 3-line file.

I've also allowed matching arbitrary template names with markdown source
names. In other words, you can provide a `my_file.temlate` that will be
prioritized over `default.html` for the file matching its name, `file.md`.
This allows quick and dirty overrides for specific pages.


Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I6a6a696472e5eda16f635e2c24b85a1513df1d16