ndg/config: add support for multiple config files and KEY=VALUE overrides
Users can now specify multiple config files via repeatable `--config-file`
flags, which are *merged in order* with later files overriding earlier ones.
This is a rename of the previous `--config` flag, which is now
repurposed as `--config KEY=VALUE` to allow fine-grained overrides
of any configuration field without editing files.


Config loading priority is: 

1. Discovered config 
2. Explicit config
3. CLI args
4. KEY=VALUE overrides via `--config`

The CLI args are set to be deprecated in a near future, so using
`--config` flag is preferable going forward.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: Ibad4e1d1b280bbd8ad3552ec549704486a6a6964