ndg/assets:rewrite custom assets copying; enable per-file minification
This is a rather large and annoying refactor that has little to no
business being a part of this PR. However, it was bound to become
necessary at *some point* so it's good that we're dealing with it.

Replaces the shitty `fs_extra::dir::copy` with walkdir-based traversal
to process each file individually. This, in turn, enables CSS and JS
minification in subdirectories at any depth, which was previously
unsupported due to low-effort bulk-copy approach.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I8f0f32e0799fba3ed354ef796df7b8ed6a6a6964