ndg/manpage: fix output issues
Fixes the "Declared by" formatting, which was one of the primary reasons
we could not use NDG for NVF. On the topic of formatting list rendering
is also fixed, and HTML wrappers are now **correctly** stripped in
defaults and examples. 

Processor re-initialization has also been fixed. This does indicate a
larger architectural issue, but that's a problem for later.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I9567ab4daf5e7f72b15f42bad07c53346a6a6964